Split the tableReader_withAssoc.cxx workflow into several workflow specs tailored to different analysis types
